+++ R/grcm39_chrlen.R
@@ -1,0 +1,14 @@
+#' @name grcm39_chrlen
+#' @aliases grcm39_chrlen
+#'
+#' @title Mouse chromosome lengths in basepairs for build GRCm39
+#'
+#' @description A vector of mouse chromosome lengths in basepairs for mouse genome build GRCm39.
+#'
+#' @source <https://www.ncbi.nlm.nih.gov/data-hub/genome/GCF_000001635.27/>
+#'
+#' @keywords datasets
+#'
+#' @examples
+#' data(grcm39_chrlen)
+NULL